rename files, test passed
